Reorganize chart inputs
