+++ web-common/src/features/canvas/inspector/chart/ChartTypeSelector.svelte
